Refactoring character card scripts
